# Conflicts:
#	mteb/benchmarks/benchmarks.py
#	mteb/evaluation/MTEB.py
#	mteb/evaluation/evaluators/RerankingEvaluator.py

+++ mteb/models/model2vec_models.py
@@ -50,7 +50,7 @@ class Model2VecWrapper(Wrapper):
         Returns:
             The encoded sentences.
         """
-        return self.static_model.encode(sentences)
+        return self.static_model.encode(sentences).astype(np.float32)
 
 
 m2v_base_glove_subword = ModelMeta(
